requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
email-validator==2.0.0.post2
gunicorn==21.2.0
